docs: New model API demo starting from a dict.
